+++ src/routes/post/[instance]/[id=integer]/+page.svelte
@@ -10,7 +10,14 @@
     import Button from '$lib/components/input/Button.svelte'
     import { page } from '$app/stores'
     import PostActions from '$lib/components/lemmy/post/PostActions.svelte'
-    import { ExclamationTriangle, Icon } from 'svelte-hero-icons'
+
+    import { 
+        ArrowPath,
+        ArrowSmallLeft,
+        ExclamationTriangle, 
+        Icon 
+    } from 'svelte-hero-icons'
+
     import Spinner from '$lib/components/ui/loader/Spinner.svelte'
     import Card from '$lib/components/ui/Card.svelte'
 
@@ -126,7 +133,31 @@
 
 <div class="flex flex-col md:flex-row gap-4 w-full">
     <div class="flex flex-col gap-2 w-full">
+        
         <div class="flex flex-col gap-3 sm:gap-4 max-w-full w-full min-w-0">
+            
+            <!--- Menu bar above post content --->
+            <div class="flex flex-row gap-4 w-full">
+                <Button class="font-normal"
+                    on:click={() => {
+                        history.back();
+                    }}
+                >
+                    <Icon src={ArrowSmallLeft} mini size="16" slot="icon" />
+                    <span class="hidden md:inline">Return to Feed</span>
+                </Button>
+
+                <Button class="font-normal"
+                    on:click={() => {
+                        reloadComments();
+                    }}
+                >
+                    <Icon src={ArrowPath} mini size="16" slot="icon" />
+                    <span class="hidden md:inline">Reload Comments</span>
+                </Button>
+            </div>
+            
+            
             {#if $page.params.instance.toLowerCase() != $instance.toLowerCase()}
             <Card cardColor="warning" class="p-4 flex flex-col gap-1">
                 <Icon
@@ -158,7 +189,9 @@
             />
   
             <h1 class="font-bold text-lg">{post.post_view.post.name}</h1>
-  
+
+
+
             {#if pType == "image"}
             <PostImage
                 name = {post.post_view.post.name}